Rename NewFeatureRequest class
This has been changed to "Changes to publishing applications or
technical advice".
