+++ app/views/changes_to_publishing_apps_requests/_request_details.html.erb
