+++ lib/support/navigation/section_groups.rb
@@ -11,7 +11,7 @@ module Support
       def initialize(current_user = nil)
         @current_user = current_user
         @groups = [
-          SectionGroup.new("Content request",sections_for(ContentAdviceRequest, ContentChangeRequest, NewFeatureRequest, UnpublishContentRequest)),
+          SectionGroup.new("Content request",sections_for(ContentAdviceRequest, ContentChangeRequest, ChangesToPublishingAppsRequest, UnpublishContentRequest)),
           SectionGroup.new("User access", sections_for(AccountsPermissionsAndTrainingRequest, RemoveUserRequest)),
           SectionGroup.new("Campaigns", sections_for(CampaignRequest)),
           SectionGroup.new("Other requests", sections_for(AnalyticsRequest, GeneralRequest, TechnicalFaultReport)),
