+++ lib/support/requests/changes_to_publishing_apps_request.rb
@@ -4,7 +4,7 @@ require 'support/gds/with_request_context'
 
 module Support
   module Requests
-    class NewFeatureRequest < Request
+    class ChangesToPublishingAppsRequest < Request
       include WithTimeConstraint
       include Support::GDS::WithRequestContext
